refactor: movendo PedidoService de pacotes
